+++ CHANGELOG.md
@@ -1,4 +1,12 @@
+## [1.0.1] - 2025-07-24
+
+### Fixed 
+- Only sails of the ship where the player is currently on or was last on can be controlled now. Last on only possible if currently not on any ship, i.e in water or on land.
+- Only ships which are owned by player can be controlled.
+
 ## [1.0.0] - 2025-07-23
 
 ### Added
 - Initial release of AutoSails mod.
+
+